chore: move receive payments under accounts settings in dashbooard navigation
