+++ src/routes/_dashboard-layout/dashboard/account/receiving-payments.tsx
@@ -187,7 +187,7 @@ function ScopeSelector({ value, scopeId, scopeIds, userPubkey, onChange }: Scope
 	)
 }
 
-export const Route = createFileRoute('/_dashboard-layout/dashboard/products/receiving-payments')({
+export const Route = createFileRoute('/_dashboard-layout/dashboard/account/receiving-payments')({
 	component: ReceivingPaymentsComponent,
 })
